remove divergent code paths for architectures
